fix(blog-post): correct time and add author to ICT.OPEN 2024 post
This commit corrects the timestamp of the blog post regarding the poster at NWO ICT.OPEN 2024 to reflect the accurate event time. Additionally, an author field has been added to provide proper credit to Niels Doorn. The inclusion of the author ensures that readers can identify the contributor responsible for this post.